# [Class] Sort our array of Class -> token references so that we can do binary instead of linear searches in it.