@spouliot 2018-10-18
I think it's fair to mention the **link all** scenario is with a very minimal app - and that anything useful would be a net gain in performance.